refactor(database): 调整数据库表结构并优化前端代码
- 将 xianyu_account 表 id 字段从 INT 修改为 BIGINT 类型
- 将 xianyu_operation_log 表 xianyu_account_id 字段从 INT 修改为 BIGINT 类型
- 移除前端自动发货记录相关 JS 文件
- 移除前端手动更新 Cookie 和 Token 相关组件代码
- 移除前端连接管理页面相关样式文件
- 在应用启动时显示访问地址提示信息

+++ src/main/resources/db/migration/V1.0.0__init_schema.sql
@@ -7,7 +7,7 @@
 -- 1. 某鱼账号表
 -- ------------------------------------------------------------
 CREATE TABLE IF NOT EXISTS xianyu_account (
-    id          INT          NOT NULL AUTO_INCREMENT PRIMARY KEY,
+    id          BIGINT       NOT NULL AUTO_INCREMENT PRIMARY KEY,
     account_note VARCHAR(100),
     unb         VARCHAR(100),
     status      TINYINT      DEFAULT 1,
